+++ package.json
@@ -12,7 +12,7 @@
   "dependencies": {
     "@emotion/react": "^11.11.3",
     "@monaco-editor/react": "^4.7.0",
-    "@nss-workshops/nss-core": "^0.1.14",
+    "@nss-workshops/nss-core": "^0.2.0",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.2.0",
     "jest": "^29.7.0",
